requirements.txt
requests
psutil
mysql-connector-python